feature: allow files with more than 2147483647 lines (#339, #341)
This commit switches internal types for line numbers to `uint64_t`. This allows to handle files with more than 2147483647 lines.

For files with smaller number of lines there should be no visible changes in behavior. For larger files there is a problem with scrolling because Qt scrollbar is limited to `int` number of steps.

+++ src/logdata/include/data/linepositionarray.h
@@ -39,6 +39,7 @@
 #ifndef LINEPOSITIONARRAY_H
 #define LINEPOSITIONARRAY_H
 
+#include <cstddef>
 #include <vector>
 
 #include "compressedlinestorage.h"
@@ -77,7 +78,7 @@ class SimpleLinePositionStorage {
     }
 
     // Element at index
-    LineOffset at( uint32_t i ) const
+    LineOffset at( size_t i ) const
     {
         return storage_.at( i );
     }
